+++ src/settings/JobProfiles/JobProfilesForm.js
@@ -111,6 +111,7 @@ export const JobProfilesFormComponent = memo(({
   const { profile } = initialValues;
   const isEditMode = Boolean(profile.id);
   const isLayerCreate = layerType === LAYER_TYPES.CREATE;
+  const isLayerDuplicate = layerType === LAYER_TYPES.DUPLICATE;
   const isSubmitDisabled = pristine || submitting;
 
   const childWrappers = useMemo(
@@ -120,9 +121,9 @@ export const JobProfilesFormComponent = memo(({
   const profileWrapperId = useMemo(
     () => {
       const lastRecord = resources.profileSnapshots?.records.at(-1);
-      return (!isLayerCreate && lastRecord) ? lastRecord?.profileWrapperId : null;
+      return (!isLayerCreate && !isLayerDuplicate && lastRecord) ? lastRecord?.profileWrapperId : null;
     },
-    [isLayerCreate, resources.profileSnapshots.records],
+    [isLayerCreate, isLayerDuplicate, resources.profileSnapshots.records],
   );
 
   const dispatch = useDispatch();
@@ -162,10 +163,43 @@ export const JobProfilesFormComponent = memo(({
   }, [childWrappers]); // eslint-disable-line react-hooks/exhaustive-deps
 
   useEffect(() => {
-    const contentData = !isLayerCreate ? childWrappers : [];
-    const getData = !isEmpty(currentJobProfileTreeContent) ? currentJobProfileTreeContent : contentData;
+    const removeKeys = (obj, keys) => {
+      let index;
+      for (const prop in obj) {
+        if (Object.hasOwn(obj, prop)) {
+          switch (typeof (obj[prop])) {
+            case 'object':
+              index = keys.indexOf(prop);
+              if (index > -1) {
+                delete obj[prop];
+              } else {
+                removeKeys(obj[prop], keys);
+              }
+              break;
+            default:
+              index = keys.indexOf(prop);
+              if (index > -1) {
+                delete obj[prop];
+              }
+              break;
+          }
+        }
+      }
+    };
 
-    setProfileTreeData(getData);
+    const contentData = !isLayerCreate ? childWrappers : [];
+    const treeData = !isEmpty(currentJobProfileTreeContent) ? currentJobProfileTreeContent : contentData;
+
+    // af far as new profileWrapperId should be created for new profiles
+    // it is needed to omit these fields during duplication because it copies
+    // data from the original job profile
+    const massageTreeData = () => (isLayerDuplicate
+      ? treeData.forEach(obj => removeKeys(obj, ['profileWrapperId']))
+      : treeData);
+
+    massageTreeData();
+
+    setProfileTreeData(treeData);
   }, [isLayerCreate, childWrappers]); // eslint-disable-line react-hooks/exhaustive-deps
 
   useEffect(() => {