+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change history for ui-data-import
 
+## [7.0.7](https://github.com/folio-org/ui-data-import/tree/v7.0.7) (2024-05-09)
+
+### Bugs fixed:
+* Changes made during duplication of "Default - Create SRS MARC Authority" job profile are not saved and affect original job profile. (UIDATIMP-1622)
+* Changes made during duplication of a job with a match profile are not saved. (UIDATIMP-1637)
+
 ## [7.0.6](https://github.com/folio-org/ui-data-import/tree/v7.0.6) (2024-05-03)
 
 ### Bugs fixed: